# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /office/approved-information:
    get:
      operationId: listApprovedInformation
      tags: [portal]
      summary: Staff-only knowledge drafts and publication history
      responses:
        "200":
          description: Office information
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ApprovedInformation"
    post:
      operationId: createApprovedInformation
      tags: [portal]
      summary: Create an unpublished draft
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovedInformationInput"
      responses:
        "201":
          description: Created draft
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApprovedInformation"
  /office/approved-information/{id}:
    patch:
      operationId: updateApprovedInformation
      tags: [portal]
      summary: Edit draft without changing the published version
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovedInformationUpdate"
      responses:
        "200":
          description: Updated draft
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApprovedInformation"
        "404":
          description: Information not found
        "409":
          description: Information changed since it was loaded
  /office/approved-information/{id}/history:
    get:
      operationId: listApprovedInformationHistory
      tags: [portal]
      summary: Staff-only publication and withdrawal history
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Past publication events, newest first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ApprovedInformationHistory"
        "404":
          description: Information not found
  /office/approved-information/{id}/review:
    post:
      operationId: reviewApprovedInformation
      tags: [portal]
      summary: Mark the current draft as reviewed
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovedInformationMutation"
      responses:
        "200":
          description: Reviewed draft
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApprovedInformation"
        "404":
          description: Information not found
        "409":
          description: Information changed since it was loaded
  /office/approved-information/{id}/publish:
    post:
      operationId: publishApprovedInformation
      tags: [portal]
      summary: Publish the reviewed draft
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovedInformationMutation"
      responses:
        "200":
          description: Published information
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApprovedInformation"
        "409":
          description: Information changed or draft has not been reviewed
  /office/approved-information/{id}/unpublish:
    post:
      operationId: unpublishApprovedInformation
      tags: [portal]
      summary: Withdraw information from the assistant
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovedInformationMutation"
      responses:
        "200":
          description: Withdrawn information
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApprovedInformation"
        "404":
          description: Information not found
        "409":
          description: Information changed since it was loaded
...
  /office/staff:
    get:
      operationId: listOfficeStaff
      tags: [portal]
      summary: List office staff (owner only)
      responses:
        "200":
          description: Office staff accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OfficeStaffMember"
        "401":
          description: Authentication required
        "403":
          description: Office owner access required
    post:
      operationId: addOfficeStaff
      tags: [portal]
      summary: Give office access to an existing account by its verified email (owner only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OfficeStaffInput"
      responses:
        "201":
          description: Added staff member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OfficeStaffMember"
        "400":
          description: Invalid email
        "401":
          description: Authentication required
        "403":
          description: Office owner access required
        "404":
          description: No account has this email verified
        "409":
          description: The account already has office access
  /office/staff/{userId}:
    delete:
      operationId: removeOfficeStaff
      tags: [portal]
      summary: Remove a staff member's office access (owner only)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 200
      responses:
        "204":
          description: Access removed
        "400":
          description: The owner account cannot be removed
        "401":
          description: Authentication required
        "403":
          description: Office owner access required
        "404":
          description: Staff member not found
  /office/audit-log:
    get:
      operationId: listOfficeAuditLog
      tags: [portal]
      summary: Latest office staff access changes, newest first (owner only)
      responses:
        "200":
          description: Up to 200 most recent audit entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuditLogEntry"
        "401":
          description: Authentication required
        "403":
          description: Office owner access required
components:
  schemas:
    OfficeStaffInput:
      type: object
      additionalProperties: false
      required: [email]
      properties:
        email: { type: string, format: email, maxLength: 254 }
    OfficeStaffMember:
      type: object
      required: [userId, email, role, addedBy, createdAt]
      properties:
        userId: { type: string }
        email: { type: ["string", "null"], format: email }
        role:
          type: string
          enum: [owner, staff]
        addedBy: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
    AuditLogEntry:
      type: object
      required: [id, actorId, actorEmail, action, targetType, targetId, details, createdAt]
      properties:
        id: { type: integer }
        actorId: { type: string }
        actorEmail: { type: ["string", "null"], format: email }
        action:
          type: string
          enum: [staff.add, staff.remove, service_request.update, inquiry.answer]
        targetType: { type: string }
        targetId: { type: string }
        details:
          type: object
          additionalProperties: true
        createdAt: { type: string, format: date-time }
...
    ApprovedInformationInput:
      type: object
      additionalProperties: false
      required: [title, content]
      properties:
        title: { type: string, minLength: 3, maxLength: 120 }
        content: { type: string, minLength: 10, maxLength: 2000 }
    ApprovedInformationUpdate:
      type: object
      additionalProperties: false
      required: [title, content, expectedUpdatedAt]
      properties:
        title: { type: string, minLength: 3, maxLength: 120 }
        content: { type: string, minLength: 10, maxLength: 2000 }
        expectedUpdatedAt: { type: string, format: date-time }
    ApprovedInformationMutation:
      type: object
      additionalProperties: false
      required: [expectedUpdatedAt]
      properties:
        expectedUpdatedAt: { type: string, format: date-time }
    ApprovedInformation:
      type: object
      required: [id, title, content, publishedTitle, publishedContent, reviewedAt, publishedReviewedAt, publishedAt, updatedAt]
      properties:
        id: { type: integer }
        title: { type: string }
        content: { type: string }
        publishedTitle: { type: ["string", "null"] }
        publishedContent: { type: ["string", "null"] }
        reviewedAt: { type: ["string", "null"], format: date-time }
        publishedReviewedAt: { type: ["string", "null"], format: date-time }
        publishedAt: { type: ["string", "null"], format: date-time }
        updatedAt: { type: string, format: date-time }
    ApprovedInformationHistory:
      type: object
      required: [id, informationId, action, title, content, reviewedAt, publishedAt, changedAt]
      properties:
        id: { type: integer }
        informationId: { type: integer }
        action: { type: string, enum: [published, withdrawn] }
        title: { type: string }
        content: { type: string }
        reviewedAt: { type: string, format: date-time }
        publishedAt: { type: string, format: date-time }
        changedAt: { type: string, format: date-time }
...