# lib/api-zod/src/generated/types/approvedInformationInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ApprovedInformationInput {
  /**
     * @minLength 3
     * @maxLength 120
     */
  title: string;
  /**
     * @minLength 10
     * @maxLength 2000
     */
  content: string;
}